+++ platzky/__init__.py
@@ -1,3 +1,3 @@
-from .platzky import Engine as Engine
-from .platzky import create_app_from_config as create_app_from_config
-from .platzky import create_engine as create_engine
+from platzky.engine import Engine as Engine
+from platzky.platzky import create_app_from_config as create_app_from_config
+from platzky.platzky import create_engine as create_engine
