+++ tests/unit_tests/plugin/fake_plugin.py
@@ -3,12 +3,14 @@ from platzky.plugin.plugin import PluginBase, PluginBaseConfig
 
 class FakePluginConfig(PluginBaseConfig):
     """Configuration for FakePlugin used in tests."""
+
     test_value: str = "default"
     optional_value: int = 42
 
 
 class FakePlugin(PluginBase[FakePluginConfig]):
     """A fake plugin implementation for testing."""
+
     config_model = FakePluginConfig
 
     def process(self, app):
